Add header in the services requests

+++ frontend-quitanda/src/services/UserService.jsx
@@ -1,40 +1,41 @@
-// import { api } from './api';
 import axios from 'axios';
 import { handleError } from '../utils/handleErrors';
 import { LocalStorageKeys } from '../enums/local-storage-keys-enum';
-import { useCallback } from 'react';
 
 const url = 'http://localhost:8000';
 
-// axios.interceptors.request.use(function (config) {
-//   const user = JSON.parse(localStorage.getItem(LocalStorageKeys.USER));
-//   if (user) {
-//     config.headers.Authorization = `Bearer ${user.token}`;
-//   }
-//   return config;
-// });
+const verifyToken = () => {
+  const localUser = JSON.parse(localStorage.getItem(LocalStorageKeys.USER));
+  if (localUser) {
+    axios.defaults.headers.common['Authorization'] = `Bearer ${localUser.token}`;
+    return true;
+  } else {
+    return false;
+  }
+};
 
+// const verifyToken = () => {
+//   var jwt = require('jsonwebtoken');
+//   try {
+//     const localUser = JSON.parse(localStorage.getItem(LocalStorageKeys.USER));
+//     if (localUser) {
+//       const token = localUser.token;
+//       jwt.verify(token, "chave_do_batman");
+//       axios.defaults.headers.common['Authorization'] = `Bearer ${token}`;
+//       return true;
+//     } else {
+//       return false;
+//     }
+//   } catch (error) {
+//     alert("Sua sessão expirou.\nÉ necessário autenticar-se novamente");
+//     return false;
+//   };
+// };
+
+// NOT AUTH ROUTES
 export async function addUser(user) {
   try {
     return await axios.post(`${url}/users`, user);
-  } catch (error) {
-    handleError(error);
-  };
-}
-
-//auth required
-export async function getUsersList() {
-  try {
-    return await axios.get(`${url}/users`);
-  } catch (error) {
-    handleError(error);
-  };
-};
-
-//auth required
-export async function getAllUsersWithPagination(sortConfig, page) {
-  try {
-    return await axios.get(`${url}/users/${sortConfig}/${page}`);
   } catch (error) {
     handleError(error);
   };
@@ -48,36 +49,9 @@ export async function getAllValidUsersWithPagination(sortConfig, page) {
   };
 };
 
-//auth required
-export async function getAllInvalidUsersWithPagination(sortConfig, page) {
-  try {
-    return await axios.get(`${url}/invalid_users/${sortConfig}/${page}`);
-  } catch (error) {
-    handleError(error);
-  };
-};
-
-//auth required
-export async function getAllUsers() {
-  try {
-    return await axios.get(`${url}/users`);
-  } catch (error) {
-    handleError(error);
-  };
-};
-
 export async function getAllValidUsers() {
   try {
     return await axios.get(`${url}/valid_users`);
-  } catch (error) {
-    handleError(error);
-  };
-};
-
-//auth required
-export async function getAllInvalidUsers() {
-  try {
-    return await axios.get(`${url}/invalid_users`);
   } catch (error) {
     handleError(error);
   };
@@ -91,31 +65,65 @@ export async function getUserByUsername(username) {
   };
 };
 
-//auth required
+// AUTH ROUTES
+export async function getAllUsers() {
+  try {
+    if (verifyToken())
+      return await axios.get(`${url}/users`);
+  } catch (error) {
+    handleError(error);
+  };
+};
+
+export async function getAllUsersWithPagination(sortConfig, page) {
+  try {
+    if (verifyToken())
+      return await axios.get(`${url}/users/${sortConfig}/${page}`);
+  } catch (error) {
+    handleError(error);
+  };
+};
+
+export async function getAllInvalidUsers() {
+  try {
+    if (verifyToken())
+      return await axios.get(`${url}/invalid_users`);
+  } catch (error) {
+    handleError(error);
+  };
+};
+
+export async function getAllInvalidUsersWithPagination(sortConfig, page) {
+  try {
+    if (verifyToken())
+      return await axios.get(`${url}/invalid_users/${sortConfig}/${page}`);
+  } catch (error) {
+    handleError(error);
+  };
+};
+
 export async function updateUserByUsername(username, userData) {
   try {
-    return await axios.patch(`${url}/user/${username}`, {
-      // headers: { 'Authorization': `Bearer ${localStorage.getItem(LocalStorageKeys.TOKEN)}` },
-      body: userData,
-    });
+    if (verifyToken())
+      return await axios.patch(`${url}/user/${username}`, userData);
   } catch (error) {
     handleError(error);
   };
 };
 
-//auth required
 export async function deleteUserByUsername(username) {
   try {
-    return await axios.delete(`${url}/users/${username}`);
+    if (verifyToken())
+      return await axios.delete(`${url}/users/${username}`);
   } catch (error) {
     handleError(error);
   };
 };
 
-//auth required
 export async function updateAccessAuthorization(username, access_authorization) {
   try {
-    return await axios.patch(`${url}/user/${username}/access_authorization`, { "access_authorization": access_authorization });
+    if (verifyToken())
+      return await axios.patch(`${url}/user/${username}/access_authorization`, { "access_authorization": access_authorization });
   } catch (error) {
     handleError(error);
   };
